fix encoding of man page links
man page links like L<printf(3)> would entity encode the man page when
generating the link, then entity encode the full link when using it.
This double encoding is very unlikely to matter, as man page names won't
include any characters that need to be entity encoded. But it's still
incorrect. When generating the link, the man page name should be URL
encoded. When adding to the output, it that link should be entity
encoded.